start with refactoring so the file is easier to understand
