+++ app/assets/main/react_components/dashboard/climate_actions/carousel/CategoryColor.jsx
